+++ src/app/_components/Issues.tsx
@@ -3,6 +3,13 @@
 import { useIssues } from '@/swr/client/issues';
 import Card from '@/components/common/organisms/Card';
 import SkeltonCard from '@/components/common/organisms/SkeltonCard';
+import { styled } from '@mui/material';
+
+const StyledIssueCard = styled('div')`
+  width: 85%;
+  height: auto;
+  margin-left: auto;
+`;
 
 type Props = {
   milestoneNumber: number;
@@ -16,7 +23,7 @@ const Issues = ({ milestoneNumber }: Props) => {
   const { data, error } = useIssues(milestoneNumber);
   if (error) return <></>;
   return (
-    <>
+    <StyledIssueCard>
       {data ? (
         data.map((item, index) => (
           <Card
@@ -29,7 +36,7 @@ const Issues = ({ milestoneNumber }: Props) => {
       ) : (
         <SkeltonCard />
       )}
-    </>
+    </StyledIssueCard>
   );
 };
